a better vibe-coded version of the server
This time I used Claude Code from scratch to build the server, using
the MBTA v3 API Swagger definition as a start, and the COTA static GTFS
and dynamic GTFS-realtime feeds as the data sources.  I'm much happier
with the setup, with the data spread out over several files and packages
in a reasonable setup.